requirements.txt
flask~=3.0.3
werkzeug~=3.0.3
jinja2~=3.1.4
paho-mqtt~=1.6.1

